feat(064): refactor game screen components
