+++ app/src/main/java/org/cescfe/numpairs/ui/components/AvailableNumberChipDefaults.kt
@@ -1,0 +1,52 @@
+package org.cescfe.numpairs.ui.components
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+internal val CHIP_MIN_HEIGHT = 48.dp
+internal val CHIP_CORNER_RADIUS = 14.dp
+internal val KNOWN_CHIP_BORDER_WIDTH = 1.dp
+internal val MODIFIABLE_CHIP_BORDER_WIDTH = 1.5.dp
+
+@Composable
+internal fun chipColorsFor(style: AvailableNumberChipStyle): AvailableNumberChipColors {
+    val colorScheme = MaterialTheme.colorScheme
+
+    return when (style) {
+        AvailableNumberChipStyle.KNOWN -> AvailableNumberChipColors(
+            containerColor = colorScheme.surface,
+            contentColor = colorScheme.onSurface,
+            border = BorderStroke(
+                width = KNOWN_CHIP_BORDER_WIDTH,
+                color = colorScheme.outline
+            )
+        )
+
+        AvailableNumberChipStyle.HIDDEN -> AvailableNumberChipColors(
+            containerColor = colorScheme.primaryContainer,
+            contentColor = colorScheme.onPrimaryContainer,
+            border = BorderStroke(
+                width = MODIFIABLE_CHIP_BORDER_WIDTH,
+                color = colorScheme.primary
+            )
+        )
+
+        AvailableNumberChipStyle.PLAYER_ENTERED -> AvailableNumberChipColors(
+            containerColor = colorScheme.secondaryContainer,
+            contentColor = colorScheme.onSecondaryContainer,
+            border = BorderStroke(
+                width = MODIFIABLE_CHIP_BORDER_WIDTH,
+                color = colorScheme.secondary
+            )
+        )
+    }
+}
+
+internal data class AvailableNumberChipColors(
+    val containerColor: Color,
+    val contentColor: Color,
+    val border: BorderStroke?
+)
